Separate aws infra create code into different files
